requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
ffmpeg-python==0.1.17
future==0.17.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-speech==0.36.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
pbr==5.1.1
pendulum==2.0.4
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pydub==0.23.0
pylint==2.2.2
pytest==4.0.2
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
pytzdata==2018.7
requests==2.21.0
robber==1.1.4
rsa==4.0
six==1.12.0
termcolor==1.1.0
typed-ast==1.1.1
urllib3==1.24.1
vibora==0.0.6
wrapt==1.10.11
